Eclipse Settings
Generally it is thought to not to version control Eclipse settings
directory. But doing so makes it difficult for a new developer to set an
Eclipse project without errors sometime in future when the old developer
who has created the project is not there/busy. Hence in this case it is
better to version control .settings directory of Eclipse.

On branch master
Your branch is up to date with 'origin/master'.

Changes to be committed:
	new file:   .settings/.jsdtscope
	new file:   .settings/org.eclipse.core.resources.prefs
	new file:   .settings/org.eclipse.jdt.core.prefs
	new file:   .settings/org.eclipse.jpt.core.prefs
	new file:   .settings/org.eclipse.m2e.core.prefs
	new file:   .settings/org.eclipse.wst.common.component
	new file:   .settings/org.eclipse.wst.common.project.facet.core.prefs.xml
	new file:   .settings/org.eclipse.wst.common.project.facet.core.xml
	new file:   .settings/org.eclipse.wst.jsdt.ui.superType.container
	new file:   .settings/org.eclipse.wst.jsdt.ui.superType.name
	new file:   .settings/org.eclipse.wst.validation.prefs
	new file:   .settings/org.hibernate.eclipse.console.prefs

 Changes not staged for commit: